Safari-609.1.11.1/LayoutTests/http/tests/security/dataURL/xss-DENIED-from-data-url-in-foreign-domain-window-open-expected.txt

CONSOLE MESSAGE: line 1: SecurityError: Blocked a frame with origin "null" from accessing a cross-origin frame. Protocols, domains, and ports must match.
ALERT: PASS: Exception thrown successfully.
Opener Frame

Pass: Cross frame access from a data: URL on a different domain was denied.
